+++ admin/contact_us.php
@@ -1,12 +1,14 @@
 <?php 
 include('top.php');
-
+if($_SESSION['ADMIN_ROLE']=='0'){
+	redirect('product');
+}
 if(isset($_GET['type']) && $_GET['type']!=='' && isset($_GET['id']) && $_GET['id']>0){
 	$type=get_safe_value($_GET['type']);
 	$id=get_safe_value($_GET['id']);
 	if($type=='delete'){
 		mysqli_query($con,"delete from contact_us where id='$id'");
-		redirect('contact_us.php');
+		redirect('contact_us');
 	}
 }
 
@@ -26,36 +28,35 @@ $res=mysqli_query($con,$sql);
                         <tr>
                             <th>#</th>
                             <th>Name</th>
-							<th>Email</th>
-							<th>Subject</th>
-							<th>Message</th>
+                            <th>Email</th>
+                            <th>Subject</th>
+                            <th>Message</th>
                             <th>Actions</th>
                         </tr>
                       </thead>
                       <tbody>
                         <?php if(mysqli_num_rows($res)>0){
-						$i=1;
-						while($row=mysqli_fetch_assoc($res)){
-						?>
-						<tr>
+                            $i=1;
+                            while($row=mysqli_fetch_assoc($res)){
+                            ?>
+                            <tr>
                             <td><?php echo $i?></td>
                             <td><?php echo $row['name']?></td>
-							<td><?php echo $row['email']?></td>
-							<td><?php echo $row['mobile']?></td>
-							<td><?php echo $row['subject']?></td>
-							<td><?php echo $row['message']?></td>
-							<td>
-								<a href="?id=<?php echo $row['id']?>&type=delete"><label class="badge badge-danger delete_red hand_cursor">Delete</label></a>
-							</td>
+                            <td><?php echo $row['email']?></td>
+                            <td><?php echo $row['subject']?></td>
+                            <td><?php echo $row['meassage']?></td>
+                            <td>
+                              <a href="?id=<?php echo $row['id']?>&type=delete"><label class="badge badge-danger delete_red hand_cursor">Delete</label></a>
+                            </td>
                            
                         </tr>
-                        <?php 
-						$i++;
-						} } else { ?>
-						<tr>
-							<td colspan="5">No data found</td>
-						</tr>
-						<?php } ?>
+                                  <?php 
+                      $i++;
+                      } } else { ?>
+                      <tr>
+                        <td colspan="5">No data found</td>
+                      </tr>
+                      <?php } ?>
                       </tbody>
                     </table>
                   </div>
